==> 01_transactions.ipynb <==
from pathlib import Path
import json
import numpy as np
import pandas as pd
from tqdm import tqdm
from scml import pandasx as pdx

percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
pd.set_option("use_inf_as_na", True)
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()

txn2id = {
    "clicks": 0,
    "carts": 1,
    "orders": 2,
}
rows = []
input_files = [Path("input/train.jsonl"), Path("input/test.jsonl")]
for fp in input_files:
    with open(str(fp)) as lines:
        for line in tqdm(lines, desc=str(fp)):
            jo = json.loads(line)
            session = int(jo["session"])
            prev = None
            for event in jo["events"]:
                curr = int(event["ts"])
                if prev is not None and curr<prev:
                    raise ValueError("event out-of-order")
                rows.append({
                    "session": session,
                    "aid": int(event["aid"]),
                    "ts": curr,
                    "type": txn2id[event["type"]]
                })
                prev = curr

df = pd.DataFrame.from_records(rows)
cols = ["session", "aid"]
df[cols] = df[cols].astype(np.int32)
cols = ["type"]
df[cols] = df[cols].astype(np.int8)
#df.sort_values(["session", "ts"], inplace=True, ignore_index=True)
df.info()

df.describe(percentiles=percentiles)

pdx.value_counts(df["type"])

%%time
assert df.notna().all(axis=None)
df.to_parquet("output/transactions.parquet", index=False)
