==> 02_sequences.ipynb <==
import json
from collections import Counter
import numpy as np
import pandas as pd
from tqdm import tqdm
from typing import Dict

percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
pd.set_option("use_inf_as_na", True)
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()

# already sorted by session, timestamp
df = pd.read_parquet("input/transactions.parquet")
df.info()

wc = Counter()
prev = None
words = []
rows = []
for t in tqdm(df.itertuples()):
    curr = int(getattr(t, "session"))
    if prev is not None and prev!=curr and len(words)!=0:
        rows.append({
            "session": prev,
            "seq": " ".join(words),
            "length": len(words),
        })
        words = []
    prev = curr
    weight = 1
    _type = int(getattr(t, "type"))
    first = "<click>"
    if _type==1:
        first = "<cart>"
        weight = 10
    elif _type==2:
        first = "<order>"
        weight = 100
    aid = str(getattr(t, "aid"))
    wc[aid]+=weight
    second = aid
    words += [first, second]
if len(words)!=0:
    rows.append({
        "session": prev,
        "seq": " ".join(words),
        "length": len(words),
    })

print(f"{len(wc)} words\n{wc.most_common(10)}")

s = pd.Series(wc.values())
s.describe(percentiles=percentiles)

vocab: Dict[str, int] = {
    "<s>": 0,
    "<pad>": 1,
    "</s>": 2,
    "<unk>": 3,
    "<mask>": 4,
    "<cls>": 5,
    "<click>": 6,
    "<cart>": 7,
    "<order>": 8,
}
_next = len(vocab)
for k,_ in wc.most_common(500_000):
    vocab[k] = _next
    _next+=1
print(f"len(vocab)={len(vocab):,}")

%%time
with open("output/vocab.json", "w") as f:
    json.dump(vocab, f)
with open("output/vocab.txt", "w") as f:
    words = [""] * len(vocab)
    for k,v in vocab.items():
        words[v] = k
    for w in words:
        f.write(f"{w}\n")

df = pd.DataFrame.from_records(rows)
cols = ["session"]
df[cols] = df[cols].astype(np.int32)
cols = ["length"]
df[cols] = df[cols].astype(np.int16)
df.info()

df.describe(percentiles=percentiles)

df.head()

%%time
assert df.notna().all(axis=None)
df.to_parquet("output/sequences.parquet", index=False)
