==> 10_graph.ipynb <==
import json
from collections import Counter
import networkx as nx
from networkx.readwrite import json_graph
import numpy as np
import pandas as pd
from tqdm import tqdm
from typing import Dict, List
import scml

tim = scml.Timer()
tim.start()
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
pd.set_option("use_inf_as_na", True)
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()

# already sorted by session, timestamp
df = pd.read_parquet("input/transactions.parquet")
#df = df.iloc[:1000]
df.info()

wc = Counter()
prev_sess = None
prev = None
edges = Counter()
types: List[str] = ["click", "cart", "order"]
for t in tqdm(df.itertuples()):
    curr_sess = int(getattr(t, "session"))
    if prev_sess is not None and prev_sess!=curr_sess:
        prev = None
    prev_sess = curr_sess
    aid = str(getattr(t, "aid"))
    _type = types[int(getattr(t, "type"))]
    curr = f"{aid}_{_type}"
    wc[aid]+=1
    if prev is not None and prev!=curr:
        edges[(prev, curr)]+=1
    prev = curr

print(f"{len(wc)} words\n{wc.most_common(10)}")

s = pd.Series(wc.values())
s.describe(percentiles=percentiles)

%%time
with open("output/vocab3.json", "w") as f:
    json.dump(wc.most_common(), f)

print(edges.most_common(10))

g = nx.DiGraph()
g.add_weighted_edges_from((u, v, w) for (u, v), w in edges.most_common())
print(g)

%%time
with open(f"output/graph.json", "w") as f:
    json.dump(json_graph.adjacency_data(g), f)

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
