==> 11_random_walk.ipynb <==
import json
import gc
import random
import networkx as nx
from networkx.readwrite import json_graph
import numpy as np
import pandas as pd
from tqdm import tqdm
from typing import Dict, List
import scml
import mylib

# number of context words in the left and right directions
m = 8
iters = 20
limit = 0
weight_min = 7
choices_max = 1000
type2id = {
    "click": 0,
    "cart": 1,
    "order": 2,
}

tim = scml.Timer()
tim.start()
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
pd.set_option("use_inf_as_na", True)
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything()

%%time
with open("input/vocab3.json") as f:
    id2label = [w for w, c in json.load(f)]
label2id = {k:v for v, k in enumerate(id2label)}
print(f"len(label2id)={len(label2id):,}")

del id2label
gc.collect()

%%time
with open("input/graph.json") as f:
    g = nx.adjacency_graph(json.load(f))
print(g)

# get neighbours

rows = []
_nodes = g.nodes
if limit>0:
    _nodes = list(_nodes)[:limit]
for center in tqdm(_nodes):
    center_word, center_type = tuple(center.split("_"))
    # going left: predecessor nodes
    for u, v, d in g.in_edges(center, data=True):
        if d["weight"]<weight_min:
            continue
        curr = u
        curr_word, curr_type = tuple(curr.split("_"))
        if curr_word!=center_word:
            rows.append({
                "center_word": label2id[center_word],
                "center_type": type2id[center_type],
                "outside_word": label2id[curr_word],
                "outside_type": type2id[curr_type],
            })
    # going right: successor nodes
    for u, v, d in g.out_edges(center, data=True):
        if d["weight"]<weight_min:
            continue
        curr = v
        curr_word, curr_type = tuple(curr.split("_"))
        if curr_word!=center_word:
            rows.append({
                "center_word": label2id[center_word],
                "center_type": type2id[center_type],
                "outside_word": label2id[curr_word],
                "outside_type": type2id[curr_type],
            })

# random walk

out_edges = {}
in_edges = {}
for i in range(iters):
    for center in tqdm(_nodes, desc=f"i={i}"):
        # going left: predecessor nodes
        center_word, center_type = tuple(center.split("_"))
        curr = center
        for _ in range(m):
            if curr not in in_edges:
                choices = []
                for u, v, d in g.in_edges(curr, data=True):
                    if len(choices)>=choices_max:
                        break
                    if d["weight"]<weight_min:
                        continue
                    choices.append((u, d["weight"]))
                in_edges[curr] = choices
            choices = in_edges[curr]
            if len(choices)==0:
                break
            curr = mylib.weighted_choice(choices)
            if curr==center:
                break
            curr_word, curr_type = tuple(curr.split("_"))
            if curr_word!=center_word:
                rows.append({
                    "center_word": label2id[center_word],
                    "center_type": type2id[center_type],
                    "outside_word": label2id[curr_word],
                    "outside_type": type2id[curr_type],
                })
        # going right: successor nodes
        curr = center
        for _ in range(m):
            if curr not in out_edges:
                choices = []
                for u, v, d in g.out_edges(curr, data=True):
                    if len(choices)>=choices_max:
                        break
                    if d["weight"]<weight_min:
                        continue
                    choices.append((v, d["weight"]))
                out_edges[curr] = choices
            choices = out_edges[curr]
            if len(choices)==0:
                break
            curr = mylib.weighted_choice(choices)
            if curr==center:
                break
            curr_word, curr_type = tuple(curr.split("_"))
            if curr_word!=center_word:
                rows.append({
                    "center_word": label2id[center_word],
                    "center_type": type2id[center_type],
                    "outside_word": label2id[curr_word],
                    "outside_type": type2id[curr_type],
                })

df = pd.DataFrame.from_records(rows)
more = len(df)
df.drop_duplicates(keep="first", inplace=True, ignore_index=True)
print(f"{more - len(df):,} rows dropped: duplicates")

cols = ["center_word", "outside_word"]
df[cols] = df[cols].astype(np.int32)
cols = ["center_type", "outside_type"]
df[cols] = df[cols].astype(np.int8)
df.info()
assert (df["center_word"]!=df["outside_word"]).all()

df.describe(percentiles=percentiles)

df.head(30)

%%time
assert df.notna().all(axis=None)
df.to_parquet("output/pairs.parquet", index=False)

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
